Minor adjustments

+++ src/main_bst.c
@@ -15,8 +15,8 @@ BinaryTree* buildTree(string** docInfo, string mode) {
     BinaryTree* tree = createBST();
 
     // variables to get the time of inserts
-    float totalInsTime = 0;
-    float meanInsTime = 0;
+    double totalInsTime = 0;
+    double meanInsTime = 0;
     int count = 0;
 
     int compCount = 0;
@@ -54,7 +54,7 @@ BinaryTree* buildTree(string** docInfo, string mode) {
 
     }
     // computing more stats
-    meanInsTime = ((float)totalInsTime)/count;
+    meanInsTime = ((double)totalInsTime)/count;
     compCountMean = compCount/count;
     clock_t end = clock();
     double totalTime = ((double)(end - start))/CLOCKS_PER_SEC;
@@ -63,7 +63,7 @@ BinaryTree* buildTree(string** docInfo, string mode) {
     // writing in the docs
     if (strcmp(mode, "stats") == 0) fprintf( insertCSV,
         /* "Total insertion time: %lf\nMean insertion time: %lf\nNumber of words: %d\nTree build time:%lf\n", */
-        "%lf,%lf,%d,%Lf,%d,%f;",
+        "%lf,%lf,%d,%lf,%d,%lf;",
         totalInsTime, meanInsTime, count, totalTime, compCount, compCountMean
     );
 
@@ -86,8 +86,8 @@ void getSearchInfo(string** docInfo, BinaryTree* tree) {
   clock_t start = clock();
 
   // for the stats
-  float totalSTime = 0;
-  float meanSTime = 0;
+  double totalSTime = 0;
+  double meanSTime = 0;
   int count = 0;
   int compCount = 0;
   double compCountMean = 0;
@@ -123,11 +123,11 @@ void getSearchInfo(string** docInfo, BinaryTree* tree) {
       docTimes[i] = totalSTime - (i != 0 ? docTimes[i-1] : 0);
       docComp[i] = compCount - (i != 0 ? docComp[i-1] : 0);
       fprintf(searchDocCSV, "%d,%lf,%lf,%d,%f,%d;",
-                i, docTimes[i], docTimes[i]/docLen, docComp[i],(float)docComp[i]/docLen, docLen);
+                i, docTimes[i], docTimes[i]/docLen, docComp[i],(double)docComp[i]/docLen, docLen);
 
   }
-  meanSTime = ((float)totalSTime/count);
-  compCountMean = (float)compCount/count;
+  meanSTime = ((double)totalSTime/count);
+  compCountMean = (double)compCount/count;
   clock_t end = clock();
   double totalTime = ((double)(end - start))/CLOCKS_PER_SEC;
   fprintf( searchCSV,
